requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
playwright
python-dotenv